update: move coordinate related utils to a folder
